% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq2screw.R
\name{dq2screw}
\alias{dq2screw}
\title{Convert a dual quaternion to a screw displacement}
\usage{
dq2screw(q)
}
\arguments{
\item{q}{A dual quaternion}
}
\value{
A list containing the rotation angle theta, rotation axis l,
translation along the screw axis d, and moment vector m of the screw axis.
}
\description{
Extract the parameters of a screw transformation from a dual quaternion
}
