% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq2tr.R
\name{dq2tr}
\alias{dq2tr}
\title{Convert a dual quaternion to a displacement}
\usage{
dq2tr(q)
}
\arguments{
\item{q}{A dual quaternion}
}
\value{
A list containing a translation vector t and a unit quaternion r
}
\description{
Converts a (unit) dual quaternion to a translation vector in R3 and a unit
quaternion specifying a rotation.
}
