% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQconj.R
\name{DQconj}
\alias{DQconj}
\title{Dual quaternion conjugation}
\usage{
DQconj(q, type = 1)
}
\arguments{
\item{q}{A numeric vector of class DualQuaternion}

\item{type}{Type of conjugation. See details.}
}
\value{
The conjugate of q
}
\description{
Performs three kinds of dual quaternion conjugation
}
\details{
Function implements three kinds of conjugation for a dual quaternion
q = x + ey.
\itemize{
 \item{type = 1: }{\code{q* = x - ey}}
 \item{type = 2: }{\code{q* = x* + ey*}}
 \item{type = 3: }{\code{q* = x* - ey*}}
}
}
