% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQinterp.R
\name{DQinterp}
\alias{DQinterp}
\title{Dual quaternion interpolation}
\usage{
DQinterp(q, p, t)
}
\arguments{
\item{q, p}{Dual quaternions}

\item{t}{Interpolation parameter between 0 and 1}
}
\value{
A dual quaternion
}
\description{
Interpolate between dual quaternions using dual linear blending (DLB)
}
\details{
Function interpolates between p and q by computing and then normalizing
the dual quaternion q(1-t) + pt
}
