% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQinverse.R
\name{DQinverse}
\alias{DQinverse}
\title{Inverse dual quaternion}
\usage{
DQinverse(q)
}
\arguments{
\item{q}{A numeric vector of class dual quaternion}
}
\value{
The inverse of q
}
\description{
Inverse dual quaternion
}
