% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQnorm.R
\name{DQnorm}
\alias{DQnorm}
\title{Dual quaternion norm}
\usage{
DQnorm(q)
}
\arguments{
\item{q}{A numeric vector of class DualQuaternion}
}
\value{
The norm of q. A dual number given by a numeric 2-vector.
}
\description{
Calculates the norm of a dual quaternion
}
