% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQnormalize.R
\name{DQnormalize}
\alias{DQnormalize}
\title{Normalize a dual quaternion}
\usage{
DQnormalize(q)
}
\arguments{
\item{q}{A numeric 4-vector corresponding to a unit quaternion.}
}
\value{
The normalized dual quaternion q
}
\description{
Normalize a dual quaternion
}
\details{
A dual quaternion q = r + .5e(rt)  is a unit dual quaternion when
when the rotation component r is a unit quaternion, and r and rt are orthogonal
as 4-vectors.
}
