% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQmulti.R
\name{\%DQ*\%}
\alias{\%DQ*\%}
\title{Dual quaternion multiplication}
\usage{
x \%DQ*\% y
}
\arguments{
\item{x, y}{Numeric vectors of class DualQuaternion}
}
\value{
A numeric vector of class DualQuaternion
}
\description{
Dual quaternion multiplication
}
