% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2rot.R
\name{q2rot}
\alias{q2rot}
\title{Get the rotation encoded by a quaternion}
\usage{
q2rot(q, warn = F)
}
\arguments{
\item{q}{A numeric 4-vector corresponding to a unit quaternion.}

\item{warn}{Logical. Return a warning if q is not a unit quaternion?}
}
\value{
A named vector \code{(theta, r1, r2, r3)} specifying the angle and axis of rotation.
}
\description{
Decomposes a unit quaternion into an angle and axis of rotation.
}
\details{
Function accepts a unit quaternion q and returns a named vector
containing the angle of rotation (theta) and a unit vector giving the axis of
rotation. By default, if q is not a unit quaternion (possibly due to numerical error),
it will be quietly normalized. If warn = T, the function will return a
warning when this takes place.
}
