% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2dq.R
\name{tr2dq}
\alias{tr2dq}
\title{Convert a displacement to a dual quaternion}
\usage{
tr2dq(t, r, warn = F)
}
\arguments{
\item{t}{A translation vector in R3}

\item{r}{A unit quaternion}

\item{warn}{Logical. Return a warning if q is not a unit quaternion?}
}
\value{
A numeric vector of length 8, of class "DualQuaternion"
}
\description{
Converts a translation vector and unit quaternion to a dual quaternion specifying
the corresponding displacement.
}
\details{
Function generates a dual quaternion encoding a translation specified
by v and rotation specified by q. By default, if q is not a unit quaternion
(possibly due to numerical error), it will be quietly normalized. If warn = T,
the function will return a warning when this takes place.
}
